bump to v1.0.0 release (#30)
* bump to v1.0.0 release

The API in this gem is stable now so we should cut a v1.0.0 release. 

I don't think i've modified the public API in downstream clients via #28, however i'd like to avoid breaking downstream apps by doing a minor version bump. To avoid that, this is a major bump and we will need to update downstream clients to test the public API updates before these are automatically included on build

* add changelog note

* remove deprecated method

* update the changelog

* revert the pre1 suffix of the version

* remove deprecated current_user spec

* use magic comment